web/requirements.txt
fastapi
uvicorn
jinja2
aiofiles
aiosqlite
python-dotenv
aiogram==3.20.0.post0
aiosqlite
